Add preliminary data collection and gestalt pattern analysis
Multiple consequent MarioAgentEvents and combined into EventRanges if the input is the same.
These are basically the states of gestalt music analysis.

Distances between the different states are calculated, and the boundaries are determined.

Tile indices are also stored to easily see the start and end tile for each gestalt.

Note: the distances are not currently calculated correctly; it's just for setting up the framework.