Add feature to display extracted gestalts on screen
The Bullet Bill sprites are drawn at the start and end tiles, with their
facing direction indicating whether they're the start or end.

The GestaltSprite is "killed" immediately, but it is made sure in
MarioWorld that the sprite remains visible

Best combined with pause and arrow keys for camera movement.