Preparations for pattern extractions
framework-selected folder contains the levels gathered from both originals and converted ones.

Added setters to serializable classes that may or may not be needed later.
This is to hopefully prevent further changes to the classes that could mess up the serialization.

The fix for AI to dodge flower enemies (Piranha Plant) is applied only if necessary.